Update Admin Management

+++ FUCommunityWeb/FuCommunityWebDataAccess/Repositories/ForumRepo.cs
@@ -79,6 +79,18 @@ namespace FuCommunityWebDataAccess.Repositories
 
             return (posts, totalItems);
         }
+        public async Task<List<Post>> GetPostsByCategory(int categoryID)
+        {
+            var query = _context.Posts.AsQueryable();
+
+            query = query.Where(post => post.CategoryID == categoryID);
+
+            var posts = await query
+                .OrderBy(post => post.PostID)
+                .ToListAsync();
+
+            return posts;
+        }
         public async Task<(List<Post> posts, int totalItems)> GetPostsAsync(int page, int pageSize, string searchString)
         {
             var query = _context.Posts.AsQueryable();
@@ -121,8 +133,18 @@ namespace FuCommunityWebDataAccess.Repositories
         public async Task DeletePost(int postId)
         {
             var post = await _context.Posts.FindAsync(postId);
-            _context.Posts.Remove(post);
-            await _context.SaveChangesAsync();
+            if (post != null)
+            {
+                _context.Posts.Remove(post);
+                await _context.SaveChangesAsync();
+            }
+        }
+
+        public async Task<List<Comment>> GetCommentsByPostID(int postId)
+        {
+            return await _context.Comments
+                .Where(c => c.PostID == postId)
+                .ToListAsync();
         }
 
         public async Task DeleteComment(int commentId)
